Updated earthexplorer.py to remove decommissioned datasets and workaround multiple dataset ids
This update is related to upstream issues yannforget/landsatxplore#45, yannforget/landsatxplore/#92 and possibly others.
First of all, I removed all collection 1 dataset and the Sentinel-2a dataset, which were decommissioned in the final weeks of 2022.
Then I stumbled into the same issues pointed by @jacquesmoati and @HanDuwol, which were related to the dataset ids being hardcoded into earthexplorer.py rather than extracted from the response to some of the queries. Unfortunately, in order to get the correct dataset ids for each scene, it is not sufficient to have Earth Explorer credentials, but it is also necessary to have access to the API (which can be done here: https://ers.cr.usgs.gov/profile/access). This is due to the fact that the download-option function needed to obtain the dataset id as shown by @HanDuwol https://github.com/yannforget/landsatxplore/issues/92#issuecomment-1437980872 is only available to approved users.
Since it would be pointless to restrict the use of this Python library to approved users, I implemented a similar workaround as the one proposed by @fkroeber here https://github.com/yannforget/landsatxplore/issues/45#issuecomment-877012329, but in a slightly different way which I believe to be more compact and elegant.