Update earthexplorer.py
Due to DATA_ PRODUCTS is a list, and when in a for loop, as long as there is one successful download, it can be assumed that it does not necessarily have to be downloaded successfully all three times. Therefore, throwing exceptions in the for loop is unscientific. I downloaded the data using this method successfully in the second round, but failed in the third round. The data can be downloaded locally, but there is a prompt of None of the archived ids succeeded! Update necessary Therefore, it is recommended to modify according to the above method. The caller can also determine whether the data download was successful by checking if the file name is empty.